Revert "Depend on libcurl-ssl-dev instead of libcurl4-openssl-dev."
This broke "rosdep install <pkg-that-uses-curl>" because it's not
possible to install a multi-provider virtual package
using "apt-get install $virt_pkg".

In order to fix the conflict between different versions of
libcurl4-*ssl-dev, rosdep would need to be made to either understand
virtual packages or allow 'a_pkg | b_pkg' rules.