Kobuki version bumps to 0.3.9-2
We have removed kobuki_softnode, as we are experiencing many problems with it on goovy. I didn't make a bloom release; I assume that just removing from here is enough. If this is not the case, feel free to drop this pull request and I will make a normal release.