Consolidate gstreamer 1.0 rosdeps
I'm cleaning up the rosdeps for gstreamer 1.0 in preparation for Kinetic
release of audio_common.

Ubuntu and Fedora package gstreamer 1.0 quite differently, and my goal
here is to provide a common set of rosdeps that work across platforms.
Fedora packages the plugins, libraries and gobject introspection data
into a single package, while Ubuntu packages these as three separate
packages. Since most users will want all of these, and it aligns with
the Fedora packaging strategy, I've combined all three under the same
rosdep key.

I've also created a gstreamer1.0 package which is an umbrella for the
tools and gobject introspection data that should come with gstreamer

I've removed the libgstreamer-**1.0 packages, because they don't exist
on Fedora and they're depended on by the more generic gstreamer packages
on Ubuntu. I've added them to the package lists for the relevant Ubuntu
packages so that it's clear to future rosdep maintainers that there are
already rosdep keys for these packages.