changing tornado debian entry to not use pip
because bloom tells me :
```
Key 'python-tornado' resolved to '['tornado']' with installer 'pip', which does not match the default installer 'apt'.
Failed to resolve python-tornado on debian:jessie with: Error running generator: The Debian generator does not support dependencies which are installed with the 'pip' installer.
```

simplifying debian rule for python-tornado