[Indigo][moveit] Remove metapackages temporarilly.
The packages listed in the change are already moved to a new devel repository, but bloom fails because they are already listed in rosdistro within the old repo. Therefore, in order to complete the move to a new devel repo, they need to be removed from rosdistro. Then bloom needs to be run shortly to keep the period where those packages in transition are not available as short as possible.

I removed the change from J and K, which can be done by bloom only (as already done for K #13809).