[test] Check if wiki.ros.org/PKG_NAME exists.
**Background**
Currently AFAIK there's no mechanism to check if a page on wiki.ros.org for the packages listed on rosdistro exist. Some/many of them are missing because of that.

**Approach**
Check if the corresponding wiki page exists (i.e. not "does not exist").

**Limitation**
At the time of this commit, the added check is commented out, as it does NOT work as intended
(with the commented-out code, test checks all listed packages and accesses wiki.ros.org, most of which result in http status code 503, probably due to web server's regulation).

Checking the packages that are ONLY included in the pull request should improve the situation, but the author couldn't implement a way to do so.