removing version from rdl entries
This is effectively a rollback of the attempted release #16562 The sourcedeb jobs are all failing because the release repo does not have the distro specific tags or branches. This is likely because the discussion [here](https://github.com/ros/rosdistro/pull/16562#issuecomment-350592013) about missing dependencies and bloom was forced to proceed without generating the specific tags.