Update the FreeBSD rosdistro information
These match the current pkg names instead of the older port names. It
is using some pip packages that need to become FreeBSD ports. This is
a chicken and egg problem, so we have to start somewhere. At least
with these changes it is possible to install ROS on FreeBSD.