Added ruby-dev as a separate key
For now it's a copy of key `ruby`, without the specializations for Ubuntu EOL distros.
In the future we would like to split ruby dependencies in run-time and build-time.

Furthermore some Orocos packages would require the `ruby-dev` key for compatibility with
the AutoProj package set (https://github.com/rock-core/package_set) used by the
Rock community. The `manifest.xml` file was originally supposed to be compatible to ROS/rosdep
and AutoProj.