Remove alpine pycryptodome key.
By convention and based on the other definitions this key expects to
resolve the pycryptodome dependency for python2.
The alpine package as the name would imply is for python3.
Since alpine only packages pycryptodome for python3 the definition
needs to be removed rather than changed to a non-existent py2 variant.