Fix libvlc rosdep key in Ubuntu/Debian
In Ubuntu Zesty and later and Debian stretch and later, the `vlc-nox`
package has been deprecated and turned into a transitional dummy package
that cannot be installed.  It has been replaced by the `vlc-bin` package.
For reference:
- https://packages.ubuntu.com/zesty/allpackages
- https://packages.debian.org/stretch/vlc-nox

This updates the `libvlc` rosdep key to install the correct packages
in the appropriate versions of Ubuntu and Debian.

Signed-off-by: P. J. Reed <preed@swri.org>