Add RHEL 8 rules from PowerTools repo (#24539)
This repo is published by RedHat, but it is not enabled by default. It
contains quite a few packages that were previously available in the
'base' or 'epel' repos for RHEL 7.

I also explicitly listed the AppStream repo. This repo is required for a
basic installation (similar to BaseOS), but I thought it best to list
it.