Copy bloom .ignored file if source distribution has one.
Somehow none of the repositories which had ignored packages tripped any
build alarms as I was using this script in testing. Support for ignored
files was added by checking if an ignored file is present for the source
distribution and if so, copy it to the dest distribution and commit the
change.