Rolling: Update source branches for ROS 2 core packages.
This commit updates source/doc versions for packages in the ROS 2
ros2.repos file to their current mainline with the exception of
ros2_tracing, where the ros2.repos file uses a tag, and Fast-RTPS where
we previously have not tracked an upstream branch. However, the rolling
nature of this release may make this a suitable place to change one or
both of those.