Add in the BSD license.
This is to make it clear what license the code in this repository
should be under.  While I was here, I also took the liberty of
adding the BSD license clause header to the files that were
clearly written by Open Robotics employees.  This included
relicensing one file from Apache 2.0 to BSD, but that was
writting completely by Tully so it should be fine.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>