Add in a script to sync source entries with release repo devel_branch. (#25908)
* Add in a script to sync source entries with release repo devel_branch.

When doing a release of a ROS package through bloom, it is possible
to set the branch in the rosdistro 'source' entry different from the
"devel_branch" that is used in the release repository.  This is
not recommended, as doing further releases will now cause bloom
to look at the "wrong" branch for tags.  The situation gets even
worse since both the rosdistro 'source' entry and the devel_branch
in the release repository can be changed by hand.

This commit adds a script to sync the "devel_branch" in the release
repository with the rosdistro 'source' branch.  This ensures that
future releases will always find the version from the correct
branch.

See the script itself for more documentation on how to run it,
and the options it has.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>