Disable embree_vendor repo in Galactic (#30353)
It appears that this package did not Bloom successfully, and tags for
neither of the supported binary platforms (Ubuntu Focal and RHEL 8) are
present.

The maintainer of this package can probably just re-release them with
a 'debinc' bump. Until then, removing the 'release' version will stop
the buildfarm from attempting to build the source package in vain.