Add python3-sip-devel RPM to python3-qt-bindings
In RHEL and previous versions of Fedora, python3-sip-devel was installed
indirectly as a dependency of one of the listed packages, but the
situation seems to have changed in recent Fedora releases and it should
now be listed explicitly.

Since it's certainly needed, I'm adding it for RHEL as well, since there
is clearly precedent for the indirect dependency disappearing.