Simplify pip skip and support nested 'packages' map
Because an os-level package list is propagated to individual OS versions
based on the config.yaml, it isn't necessary to detect and skip pip at
that level because the individual OS versions need to have the detection
anyway.

Also support the valid (though pedantic) YAML structure where the
package list is specified in a 'packages' sub-key.