Update release repositories for ROS 2 releases from tier4.
Responding to the request in https://github.com/ros/rosdistro/pull/31289#issuecomment-1006495518
This updates the release repository url for ROS 2 distros to all use the
ros2-gbp copy of the repository.

Since the same osqp_vendor-release repository is used for ROS Noetic as
well we could either update that url and use the ros2-gbp repository for
all distributions or maintain the existing release repository just for
ROS Noetic.