[migration-tools] Set stdin to /dev/null for git-bloom-release.
Bloom's non-interactive flags don't penetrate into the deeper
subcommands run by bloom internally. Work is being done to adapt bloom
so that it better handles these non-interactive contexts but rely on
stdin not being an interactive terminal. By default these subprocesses
are inheriting stdin from the parent script but for git bloom-release in
particular we want stdin to be null.