add move_base_swp to index
A drop-in replacement for move_base, designed to work with all existing
plugins (e.g. DWA planner, navfn, costmap2d) from the Navigation package.
The interface is fully compatible with move_base and will accept single-pose
goals and behave just as regular move_base with a couple of improvements
listed below. The new interface accepts the sparse-waypoint goals which
specifies a list of poses, with last pose being the final goal. If a
single-waypoint goal is provided the goal reduces to a final goal with no
waypoints only issued over the new interface. All other components will work,
so move_base_swp is simply a better move_base.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>