Drop release stanza for performance_test.
This package has a pip dependency preventing bloom from creating
release data for Ubuntu Jammy, which is currently a required platform
for Rolling. With the repository configured as-is, sourcepkg jobs are
currently running and failing since they cannot find the tags required
to create platform-specific source packages.

Removing the release stanza stops the attempted source package creation
and indicates that the package is not currently releasable for Rolling's
target platforms. If a future version of the package becomes
releasable, running bloom again will repopulate the stanza.