Account for supported platforms in rule minimization
The supported version list in the repo check config can be used as
context when minimizing rules to account for scenarios where not all
supported platforms use the same value.

It can also be used to formulate a forward-looking rule using the
wildcard syntax, which is standard practice in rosdep.