Make rosdep suggestions based on python?dist(foo) (#34774)
In Debian/Ubuntu, the recommended naming practice for Python packages is
to prefix with `python`, then the major Python version and a hyphen, and
then the module name (as permitted by other naming policies).

Fedora/EPEL plays more loosely with the package names, but provides a
virtual package `python?dist(foo)` for each Python module provided by a
package. We can use this virtual package to make rule suggestions for
the more uniquely named Python packages.