Use mirrorlist files for checking RPM repositories
There are three common ways to locate RPM metadata:
1. Base URL: A plain URL to the repository base
2. Mirror List: A URL to a plaintext file which contains a list of base
   URLs
3. Metalink: A URL to an XML file which contains hashes of the
   repomd.xml and various URLs to access it.

We're currently using (1). Option (3) is the way that modern Fedora
instances find repositories, but isn't supported by CentOS or AlmaLinux.
We'd also need to change some of the existing logic in our enumeration
code to support that level of hash verification.

Option (2) is fairly trivial to implement and should make our metadata
downloads faster and more reliable.