Update release_repo_url value if the release track has one.
This is an optional configuration option for individual release tracks
which is terribly common since most projects get and use the release
repo url from the ros/rosdistro distribution.yaml file.
The earlier versions of the migration tool did not check for this value
(at the time I wrote this migration script I did not even know that such
a config setting was possible).

This change should help avoid confusion in the future due to
discrepancies between the release repo url in tracks.yaml and in the
rosdistro distribution.yaml.

I would consider eliminating this value entirely rather than rewriting
it as well, which personally I like better.