Remove python key from osx entirely.
This key implicitly resolves python2, which is no longer available on
macOS via homebrew or otherwise.

Removing it will produce failures to resolve this key on that platform,
which is an intended and desired result.

The empty definition which is removed here essentially means that this
dependency is resolved by installing no packages rather than this
dependency has no definition. This would work since macOS maintained
a system installation of python2.