add ros2_robotiq_gripper src entries
Please add this package to be indexed in the rosdistro for Humble, Iron and Rolling.

Repository:
https://github.com/PickNikRobotics/ros2_robotiq_gripper

Packages:

- robotiq_controllers
- robotiq_description
- robotiq_driver

In ROS 1 there was a ros-industrial/robotiq package here:
https://github.com/ros-industrial/robotiq
However as of 2021-05-28 it is marked unmaintained.
It has not been released since Kinetic.

This is new repository built on top of ROS 2 Control.

https://github.com/PickNikRobotics/ros2_robotiq_gripper/blob/main/LICENSE

Robotiq is currently not supporting this package.
PickNik Robotics will be maintaining it.
Kinova Robotics will depend on this.

This is related to #37856 as we're preparing to release ROS 2 support for Kinova

Signed-off-by: Alex Moriarty <alex.moriarty@picknik.ai>