Add in a set of pyside2 rosdep keys. (#38142)
* Add in a set of pyside2 rosdep keys.

These keys are the necessary set to actually use the
pyside2 bindings that are available in rqt.

I'll note one oddity, which is the 'libshiboken2-dev' key
living in the rosdep/python.yaml file.  While that is
indeed a C-style library, it's main goal in life is to
bridge the C++ Qt library and Python.  So I felt like
it was appropriate in the python.yaml file.  I could
also be convinced to move it to base.yaml.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>