Use ros2-gbp release repository for twu_geometry.
This package was previously copied into the ros2-gbp repository during the Iron branching but not used for subsequent releases.

As a result, there is divergent bloom info between the ros2-gbp repository and the upstream one. But all tags, which are the important artifacts for release integrity, have been copied over to the ros2-gbp repository.

Future releases in all ROS 2 distributions should use the ros2-gbp repository which is configured in ros2-gbp/ros2-gbp-github-org#243