Use ros2-gbp release repository for pal_statistics.
The repository has been mirrored into ros2-gbp here: ros2-gbp/ros2-gbp-github-org#460.

I did not change the release repository for noetic at this time since it is not required for ROS 1 distributions, but if the maintainer prefers to use a single release repository for all distributions that is perfectly acceptable.