Update the ncurses key on Ubuntu and Debian.
The situation with these packages is complicated, but to whit:

1. There are packages named libncurses-dev, libncurses5-dev, and
libncursesw5-dev.  However, they all package exactly the same
software, namely version 6(!) of the development libraries of ncurses.

2. There are packages named libncurses5, libncurses6 and libncursesw6.
However, they all package exactly the same software, version 6(1) of the
runtime libraries of ncurses.

We change both Ubuntu and Debian to use the canonical names of these
libraries, namely libncurses-dev (for development) and libncurses6
(for runtime).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>