Fork the sophus repositories.
We've been waiting for almost a year to get some fixes
to sophus merged so that we can get turtlebot2 released
into Humble and Iron: https://github.com/stonier/sophus/pull/23

For now, use a fork under my username to do the release.
I still haven't merged that fix in, but once we switch
this over I will have the power to do so.

While we are in here, switch the source and doc branches to
be correct (we are releasing from the 1.3.x branch nowadays).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>