Update distribution.yaml
Modified key qb_device like it was in Noetic and added source... I know I may sound crazy, but now the pull request is how it was supposed to be using bloom. I really don't get why there was no source and the name was taken from the repo. I will be more careful next time.