Remove the octomap package from Rolling.
This is essentially a vendor package; upstream is not a ROS
package, even though it has some things in it to interoperate
with ROS.

The problem is that it is *also* packaged in debian as a system
package.  In Ubuntu 24.04, the system version of the package has
a different ABI than this "vendored" version.

So it turns out that we don't need this version at all.  All downstream
sources have been updated to use the system version of the package,
rather than this vendor.  All but Moveit have done releases into
Rolling as of this writing, so everything should continue to work
with the exception of Moveit.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>