Run rosdistro-reviewer in target repository context (#43606)
When using the `pull_request` event, we can't use `write` permissions on
the target repository, probably because the submitter could change the
action as part of their PR.

The `pull_request_target` event runs the action with the target
repository's state instead of the PR's merged state. This will make
testing changes to the automation a little more difficult, but also
means that the automation no longer requires approval to run for
first-time submitters.

Because the checkout operation now fetches the PR target's HEAD, we need
to explicitly fetch the PR's HEAD (but not check it out) so that the
automation can inspect the git tree to find the changes.