Remove support for copying repositories into the release org. (#45268)
This support was helpful to get started but creates problems with the
current terraform-managed ros2-gbp org and with encouraging split-brain
releases where some are made from external release repositories and some
are made from the official release org.

Before running the migration, check the source distribution for release
repository urls outside the release org and mirror them to the release
org before continuing.

The rosdistro-bloom rewrite of this script intends to support this
workflow with a separate subcommand for mirroring repos.