Fix wxGTK package name for Fedora and RHEL
When wxGTK v3 was introduced in Fedora, there was an attempt to switch
the package name from wxGTK to wxGTK3. I'm fuzzy on the details, but it
appears that we've reverted back to wxGTK moving forward.

For RHEL 9 both packages are available, but the 'wxGTK' package appears
more up-to-date. This leaves only RHEL 8 using the older package name.